Welcomer added, changed structure for simultaneous pingPong
- Welcomer protocol is added as a state based protocol. Checks the
client version and the game version
- Ping pong was moved to a private class inside the ClientThread class
due to override of the input and output streams. Now pingPong sends the
`PING` command and commandParser returns a Response once the `PONG` is
received. When the response is received, the clientThread cancels the
timeout timer and the clientThread continues with the next command.